% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_trial.R
\name{run_trial}
\alias{run_trial}
\title{run_trial}
\usage{
run_trial(
  n,
  mean_x,
  mean_y,
  sd_x,
  sd_y,
  test,
  alpha = 0.05,
  n_sim,
  sim_alpha = TRUE
)
}
\arguments{
\item{n}{sample sizes in both groups (assumed to be equal)}

\item{mean_x, mean_y}{means of both groups}

\item{sd_x, sd_y}{standard deviations of both groups}

\item{test}{type of test: "t.test", "wilcox.test" or "z.test"}

\item{alpha}{significance level (default is 0.05)}

\item{n_sim}{number of simulation repetitions.}

\item{sim_alpha}{logical: whether the type 1 error must be simulated (default is TRUE)}
}
\value{
List with power and type 1 error
}
\description{
simulation function that calculates the power and type 1 error of a particular set
of assumptions (sample sizes, means, standard deviations).
}
