% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_test.R
\name{z_test}
\alias{z_test}
\title{z_test}
\usage{
z_test(x, y, sd = 1, delta = 0)
}
\arguments{
\item{x, y}{vectors of observations}

\item{sd}{an assumed common standard deviation}

\item{delta}{the mean difference corresponding to the null hypothesis}
}
\value{
List with z-statistics and p-value
}
\description{
Two-Sample z-test for Comparing Two Means
}
